requirements.txt
streamlit==1.35.0
duckdb==0.10.3
pandas==2.2.2
plotly==5.22.0
requests==2.32.3
python-dotenv==1.0.1
numpy==1.26.4
pyarrow==16.1.0
